requirements.txt
PyYaml==6.0.2
vertexai==1.71.1
streamlit==1.44.1
ffmpeg-python==0.2.0
# sudo apt-get install ffmpeg -y
google-cloud-storage==2.19.0
google-genai==1.16.1
